bgpd: Fix compilation for Debian 11 when printing uint64 values
build	25-Mar-2026 08:53:54	  375 | #define PRIu64 "Lu"
build	25-Mar-2026 08:53:54	bgpd/bgp_flowspec_util.c: In function ‘bgp_flowspec_bitmask_decode’:
build	25-Mar-2026 08:53:54	bgpd/bgp_flowspec_util.c:476:44: error: format ‘%Lu’ expects argument of type ‘long long unsigned int’, but argument 4 has type ‘uint64_t’ {aka ‘long unsigned int’} [-Werror=format=]
build	25-Mar-2026 08:53:54	  476 |    len_written = snprintf(ptr, len_string, "%" PRIu64, value);
build	25-Mar-2026 08:53:54	      |                                            ^~~         ~~~~~
build	25-Mar-2026 08:53:54	      |                                                        |
build	25-Mar-2026 08:53:54	      |                                                        uint64_t {aka long unsigned int}
build	25-Mar-2026 08:53:54	In file included from ./lib/zebra.h:13,
build	25-Mar-2026 08:53:54	                 from bgpd/bgp_flowspec_util.c:8:
build	25-Mar-2026 08:53:54	./lib/compiler.h:375:18: note: format string is defined here

+++ bgpd/bgp_flowspec_util.c
@@ -336,7 +336,8 @@ int bgp_flowspec_op_decode(enum bgp_flowspec_util_nlri_t type,
 					ptr += len_written;
 				}
 			}
-			len_written = snprintf(ptr, len_string, " %" PRIu64 " ", value);
+			len_written = snprintf(ptr, len_string, " %" PRIu64 " ",
+					       (unsigned long long)value);
 			if (len_written > 0 && len_written < len_string) {
 				len_string -= len_written;
 				ptr += len_written;
@@ -473,7 +474,8 @@ int bgp_flowspec_bitmask_decode(enum bgp_flowspec_util_nlri_t type,
 					ptr += len_written;
 				}
 			}
-			len_written = snprintf(ptr, len_string, "%" PRIu64, value);
+			len_written = snprintf(ptr, len_string, "%" PRIu64,
+					       (unsigned long long)value);
 			if (len_written > 0 && len_written < len_string) {
 				len_string -= len_written;
 				ptr += len_written;
